+++ app/routes/__layout/links/new.tsx
@@ -7,10 +7,12 @@ import { createLink } from "~/models/link.server";
 import { requireUserId } from "~/session.server";
 import { useEventCallback } from "~/util/analytics";
 import { validateFormData } from "~/util/linkFormData.server";
+import { decodeStringArray } from "~/util/stringArray";
 
 type LoaderData = {
   url?: string;
   description?: string;
+  tags?: string[];
 };
 
 export const loader: LoaderFunction = async ({ request }) => {
@@ -20,17 +22,21 @@ export const loader: LoaderFunction = async ({ request }) => {
   const url = reqUrl.searchParams.get("url") ?? undefined;
   const title = reqUrl.searchParams.get("title") ?? undefined;
   const description = reqUrl.searchParams.get("description") ?? undefined;
+  const tagsRaw = reqUrl.searchParams.get("tags") ?? undefined;
+  const tags = tagsRaw
+    ? Array.from(new Set(decodeStringArray(tagsRaw)))
+    : undefined;
 
   if (!url && description) {
     // Android doesn't send URLs, because its intents don't support it.
     const match = description.match(/^(?:([\S\s]*)\s+)?(https?:\/\/[^\s]+)$/);
     if (match) {
       const [, desc, link] = match;
-      return json<LoaderData>({ url: link, description: desc ?? title });
+      return json<LoaderData>({ url: link, description: desc ?? title, tags });
     }
   }
 
-  return json<LoaderData>({ url, description });
+  return json<LoaderData>({ url, description, tags });
 };
 
 interface ActionData {
@@ -74,7 +80,7 @@ export default function NewLinkPage() {
           initialValues={{
             url: data.url ?? "",
             description: data.description ?? "",
-            tags: [],
+            tags: data.tags ?? [],
           }}
           onSubmit={useEventCallback({
             name: "create-link",