+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- ### Fixed -->
 <!-- ### Security -->
 
-## [Unreleased](https://github.com/s-thom/linkdrop/compare/v1.5.1...HEAD) - DATE
+## [Unreleased](https://github.com/s-thom/linkdrop/compare/v1.6.0...HEAD) - DATE
+
+## [1.6.0](https://github.com/s-thom/linkdrop/compare/v1.5.1...v1.6.0) - 2022-10-23
 
 ### Added